TBBTScriptExtractor/requirements.txt
requests
beautifulsoup4
fpdf2
pyppeteer
pdfkit
